requirements.txt
astor==0.8.1
attrs==22.2.0
bandit==1.7.5
certifi==2022.12.7
charset-normalizer==3.1.0
darglint==1.8.1
docutils==0.19
eradicate==2.2.0
exceptiongroup==1.1.1
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.12.6
flake8-commas==2.1.0
flake8-comprehensions==3.11.1
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
iniconfig==2.0.0
isort==5.12.0
markdown-it-py==2.2.0
mccabe==0.6.1
mdurl==0.1.2
packaging==23.0
pbr==5.11.1
pep8-naming==0.13.2
pluggy==1.0.0
pycodestyle==2.8.0
pydocstyle==6.3.0
pyflakes==2.4.0
Pygments==2.14.0
pytest==7.2.2
pytest-asyncio==0.21.0
pytest-mock==3.10.0
python-decouple==3.8
PyYAML==6.0
requests==2.28.2
responses==0.23.1
restructuredtext-lint==1.4.0
rich==13.3.3
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==5.0.0
tomli==2.0.1
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
urllib3==1.26.15
wemake-python-styleguide==0.17.0
